requirements.txt
agavepy==0.9.5
ansible-cmdb==1.31
asgiref==3.4.1
attrs==21.2.0
backports.ssl-match-hostname==3.7.0.1
bokeh==2.3.3
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.9
cheroot==8.5.2
CherryPy==18.6.1
cloudpickle==2.0.0
configparser==5.2.0
cryptography==40.0.2
cycler==0.11.0
Django==3.1.14
future==0.18.2
gitdb==4.0.9
GitPython==3.1.20
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
jaraco.classes==3.2.1
jaraco.collections==3.4.0
jaraco.functools==3.4.0
jaraco.text==3.6.0
Jinja2==3.0.3
jsonxs==0.6
kiwisolver==1.3.1
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.3.4
more-itertools==8.12.0
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pgcopy==1.5.0
pika==1.2.0
Pillow==8.4.0
pip==21.3.1
pluggy==1.0.0
portend==3.0.0
protobuf==3.19.1
psycopg2==2.9.2
psycopg2-binary==2.9.2
py==1.11.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-hostlist==1.21
python-memcached==1.59
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
setuptools==39.2.0
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
tempora==4.1.2
termcolor==1.1.0
toml==0.10.2
tornado==6.1
typing==3.7.4.3
typing_extensions==4.0.1
urllib3==1.26.7
ushlex==0.99.1
websocket-client==0.53.0
zc.lockfile==2.0
zipp==3.6.0
gunicorn
